Made access_ref_count_ be an intptr_t.

+++ src/buffer_cache/alt/alt.hpp
@@ -214,8 +214,8 @@ private:
     alt_snapshot_node_t *snapshot_node_;
 
     // Keeps track of how many alt_buf_{read|write}_t have been created for
-    // this lock.
-    int access_ref_count_;
+    // this lock, for assertion/guarantee purposes.
+    intptr_t access_ref_count_;
 
     DISABLE_COPYING(alt_buf_lock_t);
 };
